ros_control: splitted hardware interface into multiple interfaces
