+++ bitbots_ros_control/include/bitbots_ros_control/imu_hardware_interface.h
@@ -1,0 +1,50 @@
+#ifndef IMU_HARWARE_INTERFACE_H
+#define IMU_HARWARE_INTERFACE_H
+
+#include <ros/ros.h>
+#include <string>
+
+#include <diagnostic_msgs/DiagnosticStatus.h>
+#include <diagnostic_msgs/DiagnosticArray.h>
+
+#include <hardware_interface/imu_sensor_interface.h>
+#include <hardware_interface/robot_hw.h>
+#include <dynamic_reconfigure/server.h>
+
+#include <bitbots_ros_control/bitbots_ros_control_paramsConfig.h>
+
+#include <dynamixel_workbench/dynamixel_driver.h>
+
+namespace bitbots_ros_control
+{
+
+class ImuHardwareInterface : public hardware_interface::RobotHW
+{
+public:
+  ImuHardwareInterface(boost::shared_ptr<DynamixelDriver> driver);
+  void reconf_callback(bitbots_ros_control::bitbots_ros_control_paramsConfig &config, uint32_t level);
+
+  bool init(ros::NodeHandle& nh);
+  bool read();
+  void write();
+
+private:
+  bool readImu();
+
+  ros::NodeHandle _nh;
+  boost::shared_ptr<DynamixelDriver> _driver;
+  hardware_interface::ImuSensorInterface _imu_interface;
+  bool _read_imu;
+
+  uint32_t _last_seq_number;
+  double* _orientation; //quaternion (x,y,z,w)
+  double* _orientation_covariance;
+  double* _angular_velocity;
+  double* _angular_velocity_covariance;
+  double* _linear_acceleration;
+  double* _linear_acceleration_covariance;
+
+  diagnostic_msgs::DiagnosticStatus _status_IMU;
+
+
+}
