+++ bitbots_ros_control/src/dynamixel_controller_manager_node.cpp
@@ -10,8 +10,25 @@ int main(int argc, char** argv)
   ros::init(argc, argv, "dynamixel_controller_manager");
   ros::NodeHandle pnh("~");
 
-  // Load dynamixels
-  bitbots_ros_control::DynamixelHardwareInterface hw;
+  // init driver
+  ROS_INFO_STREAM("Loading parameters from namespace " << nh.getNamespace());
+  std::string port_name;
+  nh.getParam("dynamixels/port_info/port_name", port_name);
+  int baudrate;
+  nh.getParam("dynamixels/port_info/baudrate", baudrate);
+  boost::shared_ptr<DynamixelDriver> driver;
+  if(!driver->init(port_name.c_str(), uint32_t(baudrate))){
+    ROS_ERROR("Error opening serial port %s", port_name.c_str());
+    speak("Error opening serial port");
+    sleep(1);
+    exit(1);
+  }
+  float protocol_version;
+  nh.getParam("dynamixels/port_info/protocol_version", protocol_version);
+  driver->setPacketHandler(protocol_version);
+
+  // create hardware interfaces
+  bitbots_ros_control::DynamixelHardwareInterface hw = bitbots_ros_control::DynamixelHardwareInterface(driver);
 
   // set the dynamic reconfigure and load standard params
   dynamic_reconfigure::Server<bitbots_ros_control::bitbots_ros_control_paramsConfig> server;
@@ -24,6 +41,8 @@ int main(int argc, char** argv)
     ROS_ERROR_STREAM("Failed to initialize hardware interface.");
     return 1;
   }
+
+  speak("ros control startup successful");
 
   // Create separate queue, because otherwise controller manager will freeze
   ros::NodeHandle nh;
